requirements.txt
datetime
mediapipe
matplotlib
opencv-python
python-dotenv
requests
pytest
nest_asyncio
pandas
SQLAlchemy
pygrabber
pandas
twilio